refactor: ProfileView 하위 컴포넌트 파일 분리
